feat: Add function to upload to trace viewer to Tracer.
TASK: IL-496
